pyproject.toml
[tool.poetry]
name = "gpt-bot"
version = "0.1.0"
description = ""
authors = ["dezhik"]
readme = "README.md"
packages = [{include = "gpt_bot"}]

[tool.poetry.dependencies]
python = "^3.8"
openai = "^0.27.5"
aiogram = "^2.25.1"
python-decouple = "^3.8"
tiktoken = "^0.4.0"
pytest = "^7.3.1"
jinja2 = "^3.1.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
